feat(dashboard-rrhh): add new components for asset management dashboard
- Implement MonthlyAssetAcquisitions.vue for displaying monthly asset acquisition trends using area charts.
- Create RRHHAssetsTable.vue to list HR assets with their details and statuses.
- Add RRHHBarChart.vue to visualize weekly trends of HR assets.
- Introduce RRHHDonutChart.vue for displaying distribution of assets by type.
- Develop SmartphonesByBrandDonut.vue to show smartphone distribution by brand.
- Implement SmartphonesByDeparmentChart.vue to visualize smartphones and chargers by department.
- Create Stats.vue to display key statistics related to HR assets.
- Add mock data in mock.ts for testing and development purposes.
- Introduce a new carousel component with Carousel.vue and related files for better UI navigation.
- Update DashboardRRHH.vue to integrate new components and improve layout.

+++ app/Http/Controllers/AssetTypeController.php
@@ -6,6 +6,8 @@ use App\Models\AssetType;
 use App\Services\AssetTypeService;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Symfony\Component\CssSelector\Exception\InternalErrorException;
+use Symfony\Component\HttpFoundation\Exception\BadRequestException;
 
 class AssetTypeController extends Controller
 {
@@ -35,6 +37,8 @@ class AssetTypeController extends Controller
                 'assetType' => null,
                 'error' => 'Error al crear el tipo de activo: ' . $e->getMessage(),
             ]);
+
+            throw new InternalErrorException('Error al crear el tipo de activo: ' . $e->getMessage());
         }
 
         return back();
@@ -58,6 +62,8 @@ class AssetTypeController extends Controller
                 'assetType' => null,
                 'error' => 'Error al actualizar el tipo de activo: ' . $e->getMessage(),
             ]);
+
+            throw new InternalErrorException('Error al actualizar el tipo de activo: ' . $e->getMessage());
         }
 
         return back();
@@ -65,6 +71,10 @@ class AssetTypeController extends Controller
 
     public function destroy(AssetType $type)
     {
+        if (!$type->is_deletable) {
+            throw new BadRequestException('Este tipo de activo no se puede eliminar');
+        }
+
         try {
             $type->delete();
             Inertia::flash([
@@ -74,6 +84,8 @@ class AssetTypeController extends Controller
             Inertia::flash([
                 'error' => 'Error al eliminar el tipo de activo: ' . $e->getMessage(),
             ]);
+
+            throw new InternalErrorException('Error al eliminar el tipo de activo: ' . $e->getMessage());
         }
         return back();
